requirements.txt
appdirs==1.4.3
Django==1.11.1
django-bootstrap==0.2.4
django-bootstrap-themes==3.3.6
django-bootstrap3==8.2.3
django-jsonfield==1.0.1
django-location-field==2.0.3
django-material==0.13.1
googlemaps==2.4.6
olefile==0.44
packaging==16.8
Pillow==4.1.1
pkg-resources==0.0.0
pyparsing==2.2.0
pytz==2017.2
qrcode==5.3
requests==2.11.1
six==1.10.0
